+++ src/main/java/com/greeni/api/activities/domain/Activity.java
@@ -1,10 +1,12 @@
 package com.greeni.api.activities.domain;
 
-import com.greeni.api.activities.enums.ActivityType;
+import com.greeni.api.activities.domain.enums.ActivityType;
 import com.greeni.api.common.base.BaseEntity;
 import com.greeni.api.profiles.domain.Profile;
+
 import jakarta.persistence.*;
 import lombok.*;
+
 import org.hibernate.annotations.DynamicInsert;
 import org.hibernate.annotations.DynamicUpdate;
 
@@ -18,18 +20,18 @@ import org.hibernate.annotations.DynamicUpdate;
 @DynamicInsert
 public class Activity extends BaseEntity {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
 
-    @Enumerated(EnumType.STRING)
-    private ActivityType activityType;
+	@Enumerated(EnumType.STRING)
+	private ActivityType activityType;
 
-    private String name;
+	private String name;
 
-    private String description;
+	private String description;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "profile_id")
-    private Profile profile;
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "profile_id")
+	private Profile profile;
 }
