+++ src/main/java/com/greeni/api/activities/domain/enums/ActivityType.java
@@ -1,4 +1,4 @@
-package com.greeni.api.activities.enums;
+package com.greeni.api.activities.domain.enums;
 
 public enum ActivityType {
 	DIARY, FIVE_QUESTION, ROLE_PLAYING, BADGE
