+++ src/main/java/com/greeni/api/badges/service/BadgeServiceImpl.java
@@ -1,6 +1,6 @@
 package com.greeni.api.badges.service;
 
-import com.greeni.api.apiPayload.exception.GeneralException;
+import com.greeni.api.apiPayload.handler.GeneralException;
 import com.greeni.api.apiPayload.status.ProfileErrorStatus;
 import com.greeni.api.badges.converter.BadgeConverter;
 import com.greeni.api.badges.dto.BadgeResponseDTO;
@@ -8,6 +8,7 @@ import com.greeni.api.profiles.domain.Profile;
 import com.greeni.api.profiles.domain.mapping.ProfileBadge;
 import com.greeni.api.profiles.repository.ProfileBadgeRepository;
 import com.greeni.api.profiles.repository.ProfileRepository;
+
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -22,24 +23,24 @@ import java.util.List;
 @Transactional(readOnly = true)
 public class BadgeServiceImpl implements BadgeService {
 
-    private final ProfileRepository profileRepository;
-    private final ProfileBadgeRepository profileBadgeRepository;
+	private final ProfileRepository profileRepository;
+	private final ProfileBadgeRepository profileBadgeRepository;
 
-    @Override
-    public BadgeResponseDTO.GetBadgeListResponse getBadgeList(Long memberId, Long profileId) {
-        // Profile 엔티티 조회
-        Profile profile = profileRepository.findById(profileId)
-                .orElseThrow(() -> new GeneralException(ProfileErrorStatus.PROFILE_NOT_FOUND));
+	@Override
+	public BadgeResponseDTO.GetBadgeListResponse getBadgeList(Long memberId, Long profileId) {
+		// Profile 엔티티 조회
+		Profile profile = profileRepository.findById(profileId)
+			.orElseThrow(() -> new GeneralException(ProfileErrorStatus.PROFILE_NOT_FOUND));
 
-        // 이 Profile이 현재 로그인한 회원의 것인지 검증
-        if (!profile.getMember().getId().equals(memberId)) {
-            throw new GeneralException(ProfileErrorStatus.UNAUTHORIZED_PROFILE_ACCESS);
-        }
+		// 이 Profile이 현재 로그인한 회원의 것인지 검증
+		if (!profile.getMember().getId().equals(memberId)) {
+			throw new GeneralException(ProfileErrorStatus.UNAUTHORIZED_PROFILE_ACCESS);
+		}
 
-        // Badge 리스트 조회
-        List<ProfileBadge> badges = profileBadgeRepository.findByProfileIdOrderByCreatedAtDesc(profileId);
+		// Badge 리스트 조회
+		List<ProfileBadge> badges = profileBadgeRepository.findByProfileIdOrderByCreatedAtDesc(profileId);
 
-        // DTO 변환 후 반환
-        return BadgeConverter.toGetBadgeListResponseDTO(badges);
-    }
+		// DTO 변환 후 반환
+		return BadgeConverter.toGetBadgeListResponseDTO(badges);
+	}
 }
