+++ src/main/java/com/greeni/api/diaries/controller/DiaryController.java
@@ -3,7 +3,7 @@ package com.greeni.api.diaries.controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.greeni.api.diaries.docs.DiaryControllerDocs;
+import com.greeni.api.diaries.controller.docs.DiaryControllerDocs;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
