+++ src/main/java/com/greeni/api/diaries/domain/enums/Emotion.java
@@ -1,4 +1,4 @@
-package com.greeni.api.diaries.enums;
+package com.greeni.api.diaries.domain.enums;
 
 public enum Emotion {
 	HAPPY, SAD, ANGRY, ANXIETY, SURPRISED;
