+++ src/main/java/com/greeni/api/diaries/domain/enums/VoiceRole.java
@@ -1,0 +1,5 @@
+package com.greeni.api.diaries.domain.enums;
+
+public enum VoiceRole {
+	CHILD, GREENI;
+}
