--- src/main/java/com/greeni/api/diaries/enums/VoiceRole.java
@@ -1,5 +1,0 @@
-package com.greeni.api.diaries.enums;
-
-public enum VoiceRole {
-	CHILD, GREENI;
-}
