+++ src/main/java/com/greeni/api/members/service/MemberServiceImpl.java
@@ -10,7 +10,7 @@ import org.springframework.mail.javamail.MimeMessageHelper;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
-import com.greeni.api.apiPayload.exception.GeneralException;
+import com.greeni.api.apiPayload.handler.GeneralException;
 import com.greeni.api.apiPayload.status.MemberErrorStatus;
 import com.greeni.api.members.converter.MemberConverter;
 import com.greeni.api.members.domain.Member;
