+++ src/main/java/com/greeni/api/profiles/controller/ProfileController.java
@@ -5,13 +5,15 @@ import com.greeni.api.profiles.dto.ProfileRequestDTO;
 import com.greeni.api.profiles.dto.ProfileResponseDTO;
 import com.greeni.api.profiles.service.ProfileService;
 import com.greeni.api.security.jwt.userDetails.CustomUserDetails;
+
 import jakarta.validation.Valid;
+
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
-import com.greeni.api.profiles.docs.ProfileControllerDocs;
+import com.greeni.api.profiles.controller.docs.ProfileControllerDocs;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -22,55 +24,56 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor
 public class ProfileController implements ProfileControllerDocs {
 
-    private final ProfileService profileService;
+	private final ProfileService profileService;
 
-    // 프로필 생성
-    @PostMapping
-    public ResponseEntity<CommonResponse<ProfileResponseDTO.CreateProfileResponse>> addProfile(
-            @RequestBody @Valid ProfileRequestDTO.CreateProfileRequest dto,
-            @AuthenticationPrincipal CustomUserDetails customUserDetails
-    ) {
-        ProfileResponseDTO.CreateProfileResponse result = profileService.createProfile(customUserDetails.getId(), dto);
-        return new ResponseEntity<>(CommonResponse.created(result), HttpStatus.CREATED);
-    }
+	// 프로필 생성
+	@PostMapping
+	public ResponseEntity<CommonResponse<ProfileResponseDTO.CreateProfileResponse>> addProfile(
+		@RequestBody @Valid ProfileRequestDTO.CreateProfileRequest dto,
+		@AuthenticationPrincipal CustomUserDetails customUserDetails
+	) {
+		ProfileResponseDTO.CreateProfileResponse result = profileService.createProfile(customUserDetails.getId(), dto);
+		return new ResponseEntity<>(CommonResponse.created(result), HttpStatus.CREATED);
+	}
 
-    // 프로필 수정
-    @PatchMapping("/{profileId}")
-    public ResponseEntity<CommonResponse<ProfileResponseDTO.UpdateProfileResponse>> changeProfile(
-            @PathVariable Long profileId,
-            @RequestBody @Valid ProfileRequestDTO.UpdateProfileRequest dto,
-            @AuthenticationPrincipal CustomUserDetails customUserDetails
-    ) {
-        ProfileResponseDTO.UpdateProfileResponse result = profileService.updateProfile(customUserDetails.getId(), profileId, dto);
-        return new ResponseEntity<>(CommonResponse.onSuccess(result), HttpStatus.OK);
-    }
+	// 프로필 수정
+	@PatchMapping("/{profileId}")
+	public ResponseEntity<CommonResponse<ProfileResponseDTO.UpdateProfileResponse>> changeProfile(
+		@PathVariable Long profileId,
+		@RequestBody @Valid ProfileRequestDTO.UpdateProfileRequest dto,
+		@AuthenticationPrincipal CustomUserDetails customUserDetails
+	) {
+		ProfileResponseDTO.UpdateProfileResponse result = profileService.updateProfile(customUserDetails.getId(),
+			profileId, dto);
+		return new ResponseEntity<>(CommonResponse.onSuccess(result), HttpStatus.OK);
+	}
 
-    // 프로필 삭제
-    @DeleteMapping("/{profileId}")
-    public ResponseEntity<CommonResponse<Void>> removeProfile(
-            @PathVariable Long profileId,
-            @AuthenticationPrincipal CustomUserDetails customUserDetails
-    ) {
-        profileService.deleteProfile(customUserDetails.getId(), profileId);
-        return new ResponseEntity<>(CommonResponse.onSuccess(null), HttpStatus.OK);
-    }
+	// 프로필 삭제
+	@DeleteMapping("/{profileId}")
+	public ResponseEntity<CommonResponse<Void>> removeProfile(
+		@PathVariable Long profileId,
+		@AuthenticationPrincipal CustomUserDetails customUserDetails
+	) {
+		profileService.deleteProfile(customUserDetails.getId(), profileId);
+		return new ResponseEntity<>(CommonResponse.onSuccess(null), HttpStatus.OK);
+	}
 
-    // 프로필 목록 조회
-    @GetMapping("/list")
-    public ResponseEntity<CommonResponse<ProfileResponseDTO.GetProfileListResponse>> findProfileList(
-            @AuthenticationPrincipal CustomUserDetails customUserDetails
-    ) {
-        ProfileResponseDTO.GetProfileListResponse result = profileService.getProfileList(customUserDetails.getId());
-        return new ResponseEntity<>(CommonResponse.onSuccess(result), HttpStatus.OK);
-    }
+	// 프로필 목록 조회
+	@GetMapping("/list")
+	public ResponseEntity<CommonResponse<ProfileResponseDTO.GetProfileListResponse>> findProfileList(
+		@AuthenticationPrincipal CustomUserDetails customUserDetails
+	) {
+		ProfileResponseDTO.GetProfileListResponse result = profileService.getProfileList(customUserDetails.getId());
+		return new ResponseEntity<>(CommonResponse.onSuccess(result), HttpStatus.OK);
+	}
 
-    // 프로필 단일 조회
-    @GetMapping("/{profileId}")
-    public ResponseEntity<CommonResponse<ProfileResponseDTO.GetProfileResponse>> findProfile(
-            @PathVariable Long profileId,
-            @AuthenticationPrincipal CustomUserDetails customUserDetails
-    ) {
-        ProfileResponseDTO.GetProfileResponse result = profileService.getProfile(customUserDetails.getId(), profileId);
-        return new ResponseEntity<>(CommonResponse.onSuccess(result), HttpStatus.OK);
-    }
+	// 프로필 단일 조회
+	@GetMapping("/{profileId}")
+	public ResponseEntity<CommonResponse<ProfileResponseDTO.GetProfileResponse>> findProfile(
+		@PathVariable Long profileId,
+		@AuthenticationPrincipal CustomUserDetails customUserDetails
+	) {
+		ProfileResponseDTO.GetProfileResponse result = profileService.getProfile(customUserDetails.getId(), profileId);
+		return new ResponseEntity<>(CommonResponse.onSuccess(result), HttpStatus.OK);
+	}
 }
